requirements.txt
flask
python-dotenv
openai==0.28.0
gunicorn